requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
jupyterlab
pillow
torch
torchvision
imblearn
shap